requirements.txt
pillow==11.1.0
pyocr==0.8.5
pypdf==5.3.0
